SH: unifying search and hotsite
