+++ smarthint/manifest.gen.ts
@@ -8,8 +8,7 @@ import * as $$$1 from "./loaders/autocomplete.ts";
 import * as $$$2 from "./loaders/banners.ts";
 import * as $$$0 from "./loaders/PLPBanners.ts";
 import * as $$$3 from "./loaders/productListingPage.ts";
-import * as $$$4 from "./loaders/searchListPage.ts";
-import * as $$$5 from "./loaders/vitrinesWithProducts.ts";
+import * as $$$4 from "./loaders/vitrinesWithProducts.ts";
 import * as $$$$$$0 from "./sections/Analytics/SmarthintTracking.tsx";
 
 const manifest = {
@@ -18,8 +17,7 @@ const manifest = {
     "smarthint/loaders/banners.ts": $$$2,
     "smarthint/loaders/PLPBanners.ts": $$$0,
     "smarthint/loaders/productListingPage.ts": $$$3,
-    "smarthint/loaders/searchListPage.ts": $$$4,
-    "smarthint/loaders/vitrinesWithProducts.ts": $$$5,
+    "smarthint/loaders/vitrinesWithProducts.ts": $$$4,
   },
   "sections": {
     "smarthint/sections/Analytics/SmarthintTracking.tsx": $$$$$$0,
